-m ' < sabrown |X| D10_03_05_10_13_13 | Remote fiona > Minimize use of REAL'

+++ trunk/source/panacea/abc/global/gloin.h
@@ -129,7 +129,7 @@ extern char
  *DOMAIN_MAP_S,
  *MESH_CURVE;
 
-extern REAL
+extern double
  mesh_botspace,
  mesh_gxmax,
  mesh_gxmin,
